Hitflash and Backup Visuals
Hitflash blueprints added, just as was went over in class, unsure it if they work or not after various attempts to find out. Backup visuals created just in case the hitflash doesn't work. (One is visually apparent, the other would still need to be attatched to the blueprint stuff where the player gets damaged)